+++ includes/rank-math-support.php
@@ -21,19 +21,6 @@ add_filter( 'rank_math/opengraph/facebook/og_description', 'do_shortcode' );
 add_filter( 'rank_math/opengraph/twitter/title', 'do_shortcode' );
 add_filter( 'rank_math/opengraph/twitter/description', 'do_shortcode' );
 
-add_filter( 'rank_math/opengraph/facebook/og_title', function( $fbog ) {
-    return do_shortcode( $fbog );
-});
-add_filter( 'rank_math/opengraph/facebook/og_description', function( $fbogdesc ) {
-    return do_shortcode( $fbogdesc );
-});
-add_filter( 'rank_math/opengraph/twitter/title', function( $twtitle ) {
-    return do_shortcode( $twtitle );
-});
-add_filter( 'rank_math/opengraph/twitter/description', function( $twdesc ) {
-    return do_shortcode( $twdesc );
-});
-
 // Process shortcodes recursively in Rank Math Schema/JSON-LD.
 add_filter( 'rank_math/json_ld', function( $data, $context ) {
     if ( is_array( $data ) ) {